requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
PyMuPDF
google-generativeai
prometheus-fastapi-instrumentator